# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: WPMgr API
  version: 0.0.0
  description: |
    Control-plane REST API for WPMgr. This is the single source of truth for
    both the Go server types and the generated TypeScript client. Real paths
    and schemas are filled in during Phase 4 (skeleton) and Phase 5 (features).
  license:
    name: AGPL-3.0-only
    url: https://www.gnu.org/licenses/agpl-3.0.txt
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/backups/{snapshotId}/sql-inspection:
    get:
      operationId: getBackupSqlInspection
      summary: Inspect the SQL dump contained in a backup snapshot
      description: |
        Returns a structured report on the SQL dump artifact of a backup
        snapshot: table inventory, row/byte estimates, charset, table prefix,
        and (when the dump looks like a WordPress install) the canonical
        siteurl/home/db_version probed from wp_options. Resolution order:
        1. If the snapshot manifest carries an agent-generated inspection
           artifact, that JSON is returned (source="agent"). This is the cheap,
           always-correct path because the agent has the SQL plaintext locally.
        2. Otherwise the control plane streams the dump artifact, parses it
           with the legacy scanner, and caches the result for subsequent calls
           (source="cp-legacy"). The first request returns 202 Accepted while
           the inspection job runs; the client polls until it gets 200.
        Requires viewer+.
      tags: [backups]
      parameters:
        - $ref: "#/components/parameters/SnapshotId"
      responses:
        "200":
          description: The parsed SQL inspection report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SqlInspection"
        "202":
          description: |
            The legacy inspection job has been enqueued; the client should poll
            the same URL until the report is ready. The Location header echoes
            the polling URL for clients that follow it transparently.
          headers:
            Location:
              description: URL to poll for the ready report.
              schema:
                type: string
        "404":
          description: Snapshot not found, or has no DB artifact to inspect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    SnapshotId:
      name: snapshotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    AgentSignature:
      type: apiKey
      in: header
      name: X-WPMgr-Signature
      description: |
        Ed25519 signed-request authentication for agent->control-plane calls. The
        agent sends four headers: X-WPMgr-Agent-Key (its base64 Ed25519 public
        key, identifying the enrolled site), X-WPMgr-Timestamp (Unix seconds),
        X-WPMgr-Nonce (a unique per-request token), and X-WPMgr-Signature (the
        base64 Ed25519 signature over the canonical message
        "METHOD\nPATH\nTIMESTAMP\nNONCE\nhex(sha256(body))"). The control plane
        rejects stale timestamps and replayed nonces. (apiKey type is the closest
        OpenAPI 3.1 primitive for this header-carried signature scheme.)
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error description.
        details:
          type: object
          additionalProperties: true
...
    SqlInspection:
      type: object
      description: |
        Structured projection of a backup snapshot's SQL dump. Produced either
        by the agent at backup time (preferred; "source": "agent") or by the
        control plane's legacy streaming parser (fallback; "source":
        "cp-legacy"). `truncated: true` indicates the parser hit its CPU/time
        budget; tables/totals may be incomplete.
      required: [schema_version, source, is_wordpress, tables, generated_at]
      properties:
        schema_version:
          type: integer
          description: Report wire-shape version; bump when the JSON shape changes.
        dump_bytes:
          type: integer
          format: int64
          description: Total plaintext bytes scanned from the dump.
        charset:
          type: string
          description: Connection-level charset from SET NAMES (e.g. utf8mb4).
        collation:
          type: string
          description: Default collation declared in CREATE TABLE trailers, if any.
        table_prefix:
          type: string
          description: |
            Most common table-name prefix (chars up to the first underscore);
            empty when no consistent prefix could be determined.
        wp_version:
          type: string
          nullable: true
          description: WordPress db_version from wp_options, when detectable.
        siteurl:
          type: string
          nullable: true
          description: WordPress siteurl from wp_options, when detectable.
        home:
          type: string
          nullable: true
          description: WordPress home from wp_options, when detectable.
        is_wordpress:
          type: boolean
          description: True when a {prefix}options table is present in the dump.
        truncated:
          type: boolean
          description: True when the parser hit its budget before finishing.
        source:
          type: string
          enum: [agent, cp-legacy]
          description: |
            "agent" — the JSON was produced by the agent at backup time and
            stored as a snapshot artifact. "cp-legacy" — the CP streamed the
            dump artifact through its own parser as a fallback.
        parser_warnings:
          type: array
          items:
            type: string
          description: Non-fatal scanner warnings (oversized rows, regex anomalies).
        tables:
          type: array
          description: Per-table inventory; row/byte estimates are best-effort.
          items:
            type: object
            required: [name, rows_estimate]
            properties:
              name:
                type: string
              rows_estimate:
                type: integer
                format: int64
              bytes_estimate:
                type: integer
                format: int64
              auto_increment:
                type: integer
                format: int64
                nullable: true
              charset:
                type: string
              has_fk:
                type: boolean
        generated_at:
          type: string
          format: date-time
          description: When the report was assembled (agent or CP wall clock).
...